+++ src/main/java/com/example/ajouevent/service/EventFacadeService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.example.ajouevent.domain.ClubEvent;
 import com.example.ajouevent.dto.EventDetailResponseDto;
 import com.example.ajouevent.dto.EventResponseDto;
+import com.example.ajouevent.dto.EventWithKeywordDto;
 import com.example.ajouevent.dto.SliceResponse;
 import com.example.ajouevent.exception.CustomErrorCode;
 import com.example.ajouevent.exception.CustomException;
@@ -24,6 +25,8 @@ public class EventFacadeService {
 	private final EventQueryService eventQueryService;
 	private final EventCommandService eventCommandService;
 	private final EventRepository eventRepository;
+	private final TopicService topicService;
+	private final KeywordService keywordService;
 
 	// 이벤트 상세 조회 및 조회수 증가
 	public EventDetailResponseDto getEventDetail(Long eventId, Principal principal, HttpServletRequest request, HttpServletResponse response) {
@@ -41,16 +44,29 @@ public class EventFacadeService {
 		return eventQueryService.getEventDetail(eventId, principal);
 	}
 
-	// 글 타입별 조회 (동아리, 학생회, 공지사항, 기타)
+	private boolean isAnonymous(String userId) {
+		return "Anonymous".equals(userId);
+	}
+
+	// 토픽 기반 글 조회
 	public SliceResponse<EventResponseDto> getEventTypeList(String type, String keyword, Pageable pageable, Principal principal) {
 		if (principal != null) {
-			eventCommandService.markTopicAsRead(type, principal.getName());
+			topicService.markTopicAsRead(type, principal.getName());
 		}
 
 		return eventQueryService.getEventTypeList(type, keyword, pageable, principal);
 	}
 
-	private boolean isAnonymous(String userId) {
-		return "Anonymous".equals(userId);
+	// 키워드 기반 글 조회
+	public SliceResponse<EventWithKeywordDto> getClubEventsByKeyword(String searchKeyword, Principal principal, Pageable pageable) {
+		if (principal == null) {
+			throw new CustomException(CustomErrorCode.LOGIN_NEEDED);
+		}
+
+		keywordService.markKeywordAsRead(searchKeyword, principal.getName());
+
+		return eventQueryService.getClubEventsByKeyword(searchKeyword, principal, pageable);
 	}
+
+
 }
